Change call patterns so rrum => rrum_helper => rrum_main
rrum has one check that _must_ be conducted prior to entering the C++ code. This check indicates that `as`, `bs`, `ag`, and `bg` are all one. Failure to implement this check before _C++_ would result in a generic Rcpp error relating to cast issues. (e.g. it's not clear)

The remainder of the checks have been moved into _C++_ and are found in the `rrum_helper()` function